% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppQuaternions.R
\name{AxisAngle}
\alias{AxisAngle}
\title{Rotation matrix from angle and axis}
\usage{
AxisAngle(axis, angle)
}
\arguments{
\item{axis}{the axis of rotation}

\item{angle}{the angle of rotation, in radians}
}
\value{
The rotation matrix.
}
\description{
Rotation matrix from rotation angle and axis.
}
\examples{
R <- AxisAngle(c(1,2,3), 2)
RtoQuaternion(R)
fromAxisAngle(c(1,2,3), 2)
}
