% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppQuaternions.R
\name{fromAxisAngle}
\alias{fromAxisAngle}
\title{Quaternion from axis and angle}
\usage{
fromAxisAngle(axis, angle)
}
\arguments{
\item{axis}{rotation axis, a vector of length 3}

\item{angle}{rotation angle in radians}
}
\value{
A (normalized) quaternion.
}
\description{
Quaternion representation of a rotation given by axis and angle.
}
\examples{
 
q <- fromAxisAngle(c(1,2,4), 2)
toAxisAngle(q)
}
