% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppQuaternions.R
\name{fromQuaternion}
\alias{fromQuaternion}
\title{Rotation matrix from quaternion}
\usage{
fromQuaternion(q)
}
\arguments{
\item{q}{quaternion}
}
\value{
The rotation matrix corresponding to \code{q}.
}
\description{
Rotation matrix corresponding to a quaternion.
}
\examples{
R <- fromQuaternion(c(1,2,3,4))
RtoQuaternion(R)
}
