% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppQuaternions.R
\name{getRotation}
\alias{getRotation}
\title{Quaternion representation of a rotation between two vectors}
\usage{
getRotation(u, v)
}
\arguments{
\item{u, v}{vectors}
}
\value{
A quaternion, whose corresponding rotation transforms \code{u} 
to \code{v}.
}
\description{
Quaternion representation of a rotation transforming a vector 
to another one
}
\examples{
u <- c(1,2,3)
v <- c(3,2,1)
q <- getRotation(u,v)
R <- fromQuaternion(q)
R \%*\% u
}
