% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppQuaternions.R
\name{qprod}
\alias{qprod}
\title{Quaternion product}
\usage{
qprod(q1, q2)
}
\arguments{
\item{q1, q2}{quaternions}
}
\value{
A quaternion, the product of \code{q1} and \code{q2}.
}
\description{
Product of two quaternions.
}
