\name{RcppQuaternions-package}
\alias{RcppQuaternions-package}
\alias{RcppQuaternions}
\docType{package}
\title{
\packageTitle{RcppQuaternions}
}
\description{
\packageDescription{RcppQuaternions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RcppQuaternions}
\packageIndices{RcppQuaternions}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RcppQuaternions}

Maintainer: \packageMaintainer{RcppQuaternions}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
