% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppQuaternions.R
\name{RtoAxisAngle}
\alias{RtoAxisAngle}
\title{Axis and angle of a rotation}
\usage{
RtoAxisAngle(R)
}
\arguments{
\item{R}{rotation matrix}
}
\value{
A list, the axis and the angle.
}
\description{
Axis and angle of a rotation given by its matrix.
}
\examples{
theta <- 2
R <- rbind(c(cos(theta), sin(theta), 0), 
           c(-sin(theta), cos(theta), 0), 
           c(0, 0, 1))
RtoAxisAngle(R)
}
