% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppQuaternions.R
\name{RtoQuaternion}
\alias{RtoQuaternion}
\title{Quaternion representation of a rotation matrix}
\usage{
RtoQuaternion(R)
}
\arguments{
\item{R}{rotation matrix}
}
\value{
The (normalized) quaternion representing \code{R}.
}
\description{
Quaternion corresponding to a rotation matrix.
}
\examples{
R <- fromQuaternion(c(1,2,3,4))
RtoQuaternion(R)
}
