% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppQuaternions.R
\name{rversor}
\alias{rversor}
\title{Random quaternions}
\usage{
rversor(n = 1)
}
\arguments{
\item{n}{positive integer, the sample size}
}
\value{
The sampled quaternions stacked in a matrix with \code{n} rows.
}
\description{
Uniform sampling of unit quaternions
}
