% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppQuaternions.R
\name{slerp}
\alias{slerp}
\title{Slerp}
\usage{
slerp(q1, q2, t)
}
\arguments{
\item{q1, q2}{quaternions}

\item{t}{numeric vector}
}
\value{
A matrix with \code{length(t)} rows, each row being an interpolated 
quaternion.
}
\description{
Interpolation of quaternions
}
\examples{
# spherical interpolation
library(rgl)
q1 <- c(0,1,0,0)
q2 <- c(0,0,1,0)
qs <- slerp(q1, q2, seq(0, 1, length.out = 10))
spheres3d(0, 0, 0, 1, color="red", alpha=0.3)
points3d(qs[,-1])
}
