% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppQuaternions.R
\name{toAxisAngle}
\alias{toAxisAngle}
\title{Axis and angle of the rotation corresponding to a quaternion}
\usage{
toAxisAngle(q)
}
\arguments{
\item{q}{quaternion}
}
\value{
A list, the axis and the angle.
}
\description{
Rotation corresponding to a quaternion, given by axis and angle.
}
\examples{
aa <- toAxisAngle(c(1,2,3,4))
fromAxisAngle(aa$axis, aa$angle)
}
